requirements.txt
Flask==3.0.0
html2docx==1.6.0
lxml==5.0.1
python_docx==1.1.0
